// src/app/shared/models/location.model.ts
export interface ILocation {
  stateName: string;
  stateId: string;
  townName?: any;
  images?: any;
  description?: any;
  hash: string;
  isRegion: boolean;
  regionName?: any;
  lastUpdatedAt: string;
  createdAt: string;
  active: boolean;
  id: number;
  averagePrice: any;
  service_image_url: any;
  service_title: any;
  service_alt: any;
}